Change how Diskstats waits for its first handful of samples.
This fixes two bugs -- One a rounding error, one a desyncronization
between the main program and the data gathering daemon.

Additionally, it changes how the daemon prints the very first
sample. Before it would unconditionally wait until the time was
%$sample-time, which meant that in a worst case scenario,
a user could be left waiting for a while before seeing a line
of output.
This commit changes that first wait to only sleep if less
than 20% away from %$sample-time.