lib/Cxn.pm: Silence a warning (in 5.14)
I think this warning actually happens everywhere, but because of
how it happens, it gets masked by older versions of Perl.

Basically, Cxn's DESTROY method checks if $self->{dbh}, and if true,
calls ->disconnect on that. However, during global destruction --
when DESTROY is called -- you have no assurances that DESTROY
will be called before the dbh was already reaped.

This commit makes sure that the $self->{dbh} we have inside DESTROY
is actually a reference.