PT-74 Fixed getSecuritySettings func
On some servers, like a mongos with config servers, this fails if session
mode is Monotonic.
On some other servers like a secondary in a replica set, this fails if the
session mode is Strong.
Added a new func to try to get this info using both modes if necessary.