Turn off statement based binlog checks
For RDS we use row based replication however `pt-table-checksum` wants
to enforce statement and despite passing in the
`--no-check-binlog-format` flag, it would still run it (even though the
documentation suggests[1] it shouldn't). This appears to be a bug
because there is another binlog format check that is wrapped in
`$o->get('check-binlog-format')` slightly further down.

To resolve this, I've updated the remaining binlog format check in the
same expression as the one below and it now honours the
`--no-check-binlog-format` as expected.

[1]: https://www.percona.com/doc/percona-toolkit/3.0/pt-table-checksum.html#cmdoption-pt-table-checksum--[no]check-binlog-format