Changes for LP1510809. ps -eaF is captured now.
Modified f argument, for F. From man pages:

" -F     Extra full format.  See the -f option, which -F implies."

The following columns have been added to the output, due to this
change:

SZ   size in physical pages of the core image of the process.
     This includes text, data, and stack space.  Device mappings
     are currently excluded; this is subject to change.

RSS  resident set size, the non-swapped physical memory that a
     task has used (inkiloBytes).

PSR  processor that process is currently assigned to.