Only set binlog_format when necessary
The documentation states that --sync-to-master or --replicate need
the binlog_format STATEMENT, but the code set the binlog format
regardless of the operation mode.

With this patch, the binlog_format is only set if --sync-to-master
or --replicate have actually been supplied on the command line,
enabling a node-to-node sync without these options without the SUPER
privilege.