Fix newlines in column names not detected
Due to mysql 5.6.40 allowing newlines in column names the previous line:
my @defs   = $ddl =~ m/^(\s+`.*?),?$/gm; 
Is broken due to it detecting newlines as line ends 'm' argument at the end. 
To correct this issue I've made use of zero-length assertions known as "lookaround" 
https://www.regular-expressions.info/lookaround.html
what does it do?
m/(?:(?<=,\n)|(?<=\(\n))(\s+`(?:.|\n)+?`.+),?/g;
TLDR:
Treat the string as one long string and don't treat \n as the end of a line.
look for (\s+`.+?`.+),?
or       (\s+`\n+?`.+),?
if one of those matches look at what preceeds the string
if it's ',\n'  or  ')\n'  the string matches. Only save what's in (\s+`(?:.|\n)+?`.+)


m/  is declaring this a matching regex.
(?:(?<=,\n)|(?<=\(\n))  This is a OR statement including to look-behind clauses. The ?: tells the enclosing parentheses to not store the result as a variable. I've put the two look-behinds in this OR statement below this line:
(?<=,)   Look behind the matched string for a comma, the comma must be there for this look behind to match.
(?<=\()  Look behind the matched string for a open parentheses, the open parentheses must be there.
Remember those last two look-behinds are in a OR statement so only one of them have to match.
\n(\s+`.*?) This is the actual match. Match newline character followed by one or more spaces followed by back-tick followed by any character any number of times, but don't be greedy and take the rest of the match into consideration.
,? match a comma that may not proceed the matched string or in other words don't save the comma.
/g    don't stop if this pattern matches keep looking for more patterns to the end of the string.