[pt-deadlock-logger] add --disable-buffering-stdout option
[Problem]

It is a common behaviour that Perl uses buffering for I/O
to reduce the number of syscalls from the point of performance.

This can be a problem, for example, when `pt-deadlock-logger` is
running as a Docker container (which means that no terminal is attached)
and the number of deadlock events are scarce.

The `print(2)`'s output does not appear until the buffer gets filled up,
which prevents users from getting deadlock events as soon as possible.

[Solution]

Add `--disable-buffering-stdout` to disable the buffering behaviour.

The buffering should not be disabled by default, because the buffering
is basically for improving the performances by optimizing the number of
syscalls for I/O. That's why the option should be enabled explicitly.

[Further Reading]

- https://perl.plover.com/FAQs/Buffering.html