Implement --chunk-size-hard-limit (#2)
This PR does two things:

Reverts previous changes to chunk-size-limit logic
Introduces chunk-size-hard-limit flag which sets a hard limit to avoids copying a chunk larger than this number of rows. It is an unconditional check that aborts a migration if an estimated chunk size exceeds the chosen number. Default value is 15K.
I've decided to implement chunk-size-hard-limit instead of patching logic of chunk-size-limit because the practice showed that setting chunk-size-limit=10 is not enough. I need to raise to larger value. It makes me uncomfortable because this flag is used in several places of the codebase. I don't know what are side effects in all of the places. Instead, chunk-size-hard-limit implements exactly what I want without any side effects.

Tested manually.