Fix for PT-1983 pt-summary missing one DIMM
This is a fix for PT-1983. dmidecode now doesn't print "Speed: UKNOWN" for
missing DIMMs. pt-summary code relies on the Speed field to be there, and when
it's missing, rows get concatenated (or rather not broken down correctly).
Adjusted the relevant code to not rely on the Speed field to be present.