Added test for the fix for PT-1983
Due to dmidecode changes in 3.2, pt-summary needs to be able to work
with memory-related info with some missing fields in the dmidecode
output. This commit adds a test with dmidecode output modified to
include new format.

In addition, fixing test that are failing due to commit c9ec74c2515bfa59f718a7ef034fe43cf83990c3