Fix for PT-1983: pt-summary has missing records for memory due to dmidecode changes (#513)
* Fix for PT-1983 pt-summary missing one DIMM

This is a fix for PT-1983. dmidecode now doesn't print "Speed: UKNOWN" for
missing DIMMs. pt-summary code relies on the Speed field to be there, and when
it's missing, rows get concatenated (or rather not broken down correctly).
Adjusted the relevant code to not rely on the Speed field to be present.

* Added test for the fix for PT-1983

Due to dmidecode changes in 3.2, pt-summary needs to be able to work
with memory-related info with some missing fields in the dmidecode
output. This commit adds a test with dmidecode output modified to
include new format.

In addition, fixing test that are failing due to commit c9ec74c2515bfa59f718a7ef034fe43cf83990c3

* PT-1983: updated changelog to reflect the bug fixed