Updated Makefile to avoid need of Git.
In our build environment, we don't have Git installed and the packages
are being built from the .tar.gz file downloaded from the repo.
Because of that, we cannot use git command to get the tags nor commit
sha. The new makefile creates a file that holds the version and commit
SHA used to build the go binaries.